fastapi-service/requirements.txt
asyncio
typing
fastapi
pydantic
openai
langchain
langchain_community
python-dotenv
uvicorn
